create cases.factories_test
