+++ backend-project/small_eod/cases/tests.py
@@ -1,4 +1,5 @@
 from django.test import TestCase
+from rest_framework.test import APIRequestFactory, force_authenticate
 
 from .factories import CaseFactory
 from .models import Case
@@ -6,56 +7,12 @@ from .serializers import CaseSerializer, CaseCountSerializer
 from ..dictionaries.factories import FeatureFactory, DictionaryFactory
 from ..generic.tests import (
     GenericViewSetMixin,
-    FactoryCreateObjectsMixin,
     ReadOnlyViewSetMixin,
 )
-from ..institutions.factories import InstitutionFactory
-from ..tags.factories import TagFactory
+from ..notes.factories import NoteFactory
 from ..tags.models import Tag
 from ..users.factories import UserFactory
 from ..users.serializers import UserSerializer
-from rest_framework.test import APIRequestFactory, force_authenticate
-from ..notes.factories import NoteFactory
-
-
-class CaseFactoryTestCase(FactoryCreateObjectsMixin, TestCase):
-    MODEL = Case
-    FACTORY = CaseFactory
-    FACTORY_COUNT = 2  # its slow
-
-    @classmethod
-    def create_factory(cls):
-        return cls.FACTORY.create(
-            audited_institutions=InstitutionFactory.create_batch(size=2),
-            responsible_users=UserFactory.create_batch(size=2),
-            notified_users=UserFactory.create_batch(size=2),
-            tags=TagFactory.create_batch(size=2),
-            features=FeatureFactory.create_batch(size=2),
-        )
-
-    def test_many_to_many(self):
-        """
-        Check if related objects are created.
-        """
-        audited_institutions = InstitutionFactory.create_batch(size=2)
-        responsible_users = UserFactory.create_batch(size=2)
-        notified_users = UserFactory.create_batch(size=2)
-        tags = TagFactory.create_batch(size=2)
-        features = FeatureFactory.create_batch(size=2)
-
-        case = self.FACTORY.create(
-            audited_institutions=audited_institutions,
-            responsible_users=responsible_users,
-            notified_users=notified_users,
-            tags=tags,
-            features=features,
-        )
-
-        self.assertCountEqual(audited_institutions, case.audited_institution.all())
-        self.assertCountEqual(responsible_users, case.responsible_user.all())
-        self.assertCountEqual(notified_users, case.notified_user.all())
-        self.assertCountEqual(tags, case.tag.all())
-        self.assertCountEqual(features, case.feature.all())
 
 
 class CaseCountSerializerTestCase(TestCase):
